Rename `ProgramId/` to `Vocabulary/`
