+++ CMakeLists.txt
@@ -32,11 +32,11 @@ add_library(Sts1CobcSw_FileSystem STATIC)
 add_library(Sts1CobcSw_FramSections INTERFACE)
 add_library(Sts1CobcSw_Hal STATIC)
 add_library(Sts1CobcSw_Outcome INTERFACE)
-add_library(Sts1CobcSw_ProgramId INTERFACE)
 add_library(Sts1CobcSw_Periphery STATIC)
 add_library(Sts1CobcSw_Serial INTERFACE)
 add_library(Sts1CobcSw_Telemetry STATIC)
 add_library(Sts1CobcSw_Utility STATIC)
+add_library(Sts1CobcSw_Vocabulary INTERFACE)
 add_program(HelloDummy)
 
 if(CMAKE_SYSTEM_NAME STREQUAL Generic)
