+++ Sts1CobcSw/Edu/Types.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 
-#include <Sts1CobcSw/ProgramId/ProgramId.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
 #include <Sts1CobcSw/Serial/Serial.hpp>
 #include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/ProgramId.hpp>
 
 #include <cstddef>
 #include <cstdint>
