+++ Sts1CobcSw/FileSystem/CMakeLists.txt
@@ -1,7 +1,7 @@
 target_sources(Sts1CobcSw_FileSystem PRIVATE LfsWrapper.cpp)
 target_link_libraries(
-    Sts1CobcSw_FileSystem PUBLIC etl::etl littlefs::littlefs Sts1CobcSw_Outcome
-                                 Sts1CobcSw_ProgramId Sts1CobcSw_Serial
+    Sts1CobcSw_FileSystem PUBLIC etl::etl littlefs::littlefs Sts1CobcSw_Outcome Sts1CobcSw_Serial
+                                 Sts1CobcSw_Vocabulary
 )
 
 if(CMAKE_SYSTEM_NAME STREQUAL Generic)
