--- Sts1CobcSw/ProgramId/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_link_libraries(Sts1CobcSw_ProgramId INTERFACE strong_type::strong_type Sts1CobcSw_Serial)
