+++ Sts1CobcSw/Telemetry/TelemetryRecord.hpp
@@ -2,9 +2,9 @@
 
 
 #include <Sts1CobcSw/Periphery/Eps.hpp>
-#include <Sts1CobcSw/ProgramId/ProgramId.hpp>
 #include <Sts1CobcSw/Serial/Serial.hpp>
 #include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/ProgramId.hpp>
 
 #include <bit>
 #include <cstddef>
