+++ Sts1CobcSw/Vocabulary/ProgramId.ipp
@@ -1,7 +1,7 @@
 #pragma once
 
 
-#include <Sts1CobcSw/ProgramId/ProgramId.hpp>
+#include <Sts1CobcSw/Vocabulary/ProgramId.hpp>
 
 
 namespace sts1cobcsw
