+++ iwyu.imp
@@ -56,7 +56,6 @@
   { include: ["\"Sts1CobcSw/Hal/Spi.ipp\"",                             "private", "<Sts1CobcSw/Hal/Spi.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Hal/Uart.ipp\"",                            "private", "<Sts1CobcSw/Hal/Uart.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/Fram.ipp\"",                      "private", "<Sts1CobcSw/Periphery/Fram.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/ProgramId/ProgramId.ipp\"",                 "private", "<Sts1CobcSw/ProgramId/ProgramId.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Byte.ipp\"",                         "private", "<Sts1CobcSw/Serial/Byte.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Serial.ipp\"",                       "private", "<Sts1CobcSw/Serial/Serial.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/ErrorDetectionAndCorrection.ipp\"", "private", "<Sts1CobcSw/Utility/ErrorDetectionAndCorrection.hpp>", "public"] },
@@ -65,6 +64,7 @@
   { include: ["\"Sts1CobcSw/Utility/Span.ipp\"",                        "private", "<Sts1CobcSw/Utility/Span.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/StringLiteral.ipp\"",               "private", "<Sts1CobcSw/Utility/StringLiteral.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/TimeTypes.ipp\"",                   "private", "<Sts1CobcSw/Utility/TimeTypes.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Vocabulary/ProgramId.ipp\"",                "private", "<Sts1CobcSw/Vocabulary/ProgramId.hpp>", "public"] },
 
   # Include all our headers with <> instead of ""
   { include: ["\"Sts1CobcSw/Edu/Edu.hpp\"",                             "public", "<Sts1CobcSw/Edu/Edu.hpp>", "public"] },
@@ -97,7 +97,6 @@
   { include: ["\"Sts1CobcSw/Periphery/FramMock.hpp\"",                  "public", "<Sts1CobcSw/Periphery/FramMock.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/Rf.hpp\"",                        "public", "<Sts1CobcSw/Periphery/Rf.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/Spis.hpp\"",                      "public", "<Sts1CobcSw/Periphery/Spis.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/ProgramId/ProgramId.hpp\"",                 "public", "<Sts1CobcSw/ProgramId/ProgramId.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Byte.hpp\"",                         "public", "<Sts1CobcSw/Serial/Byte.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Serial.hpp\"",                       "public", "<Sts1CobcSw/Serial/Serial.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Telemetry/TelemetryMemory.hpp\"",           "public", "<Sts1CobcSw/Telemetry/TelemetryMemory.hpp>", "public"] },
@@ -112,6 +111,8 @@
   { include: ["\"Sts1CobcSw/Utility/Span.hpp\"",                        "public", "<Sts1CobcSw/Utility/Span.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/StringLiteral.hpp\"",               "public", "<Sts1CobcSw/Utility/StringLiteral.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/TimeTypes.hpp\"",                   "public", "<Sts1CobcSw/Utility/TimeTypes.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Vocabulary/ProgramId.hpp\"",                "public", "<Sts1CobcSw/Vocabulary/ProgramId.hpp>", "public"] },
+  # TODO: Update the following mappings which are now in CobcSoftware/
   { include: ["\"Sts1CobcSw/CommandParser.hpp\"",                       "public", "<Sts1CobcSw/CommandParser.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/EduCommunicationErrorThread.hpp\"",         "public", "<Sts1CobcSw/EduCommunicationErrorThread.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/EduListenerThread.hpp\"",                   "public", "<Sts1CobcSw/EduListenerThread.hpp>", "public"] },
